fix: show country for country code
show country for the country code get from api and show it in profile page
