+++ src/app/constants/core-api.ts
@@ -1,1002 +1,3 @@
 export enum Crisis {
   COVID19 = 1,
 }
-
-export const countryList = [
-  {
-    name: 'Afghanistan',
-    isoAlphaTwoCode: 'AF',
-  },
-  {
-    name: 'Åland Islands',
-    isoAlphaTwoCode: 'AX',
-  },
-  {
-    name: 'Albania',
-    isoAlphaTwoCode: 'AL',
-  },
-  {
-    name: 'Algeria',
-    isoAlphaTwoCode: 'DZ',
-  },
-  {
-    name: 'American Samoa',
-    isoAlphaTwoCode: 'AS',
-  },
-  {
-    name: 'Andorra',
-    isoAlphaTwoCode: 'AD',
-  },
-  {
-    name: 'Angola',
-    isoAlphaTwoCode: 'AO',
-  },
-  {
-    name: 'Anguilla',
-    isoAlphaTwoCode: 'AI',
-  },
-  {
-    name: 'Antarctica',
-    isoAlphaTwoCode: 'AQ',
-  },
-  {
-    name: 'Antigua and Barbuda',
-    isoAlphaTwoCode: 'AG',
-  },
-  {
-    name: 'Argentina',
-    isoAlphaTwoCode: 'AR',
-  },
-  {
-    name: 'Armenia',
-    isoAlphaTwoCode: 'AM',
-  },
-  {
-    name: 'Aruba',
-    isoAlphaTwoCode: 'AW',
-  },
-  {
-    name: 'Australia',
-    isoAlphaTwoCode: 'AU',
-  },
-  {
-    name: 'Austria',
-    isoAlphaTwoCode: 'AT',
-  },
-  {
-    name: 'Azerbaijan',
-    isoAlphaTwoCode: 'AZ',
-  },
-  {
-    name: 'Bahamas',
-    isoAlphaTwoCode: 'BS',
-  },
-  {
-    name: 'Bahrain',
-    isoAlphaTwoCode: 'BH',
-  },
-  {
-    name: 'Bangladesh',
-    isoAlphaTwoCode: 'BD',
-  },
-  {
-    name: 'Barbados',
-    isoAlphaTwoCode: 'BB',
-  },
-  {
-    name: 'Belarus',
-    isoAlphaTwoCode: 'BY',
-  },
-  {
-    name: 'Belgium',
-    isoAlphaTwoCode: 'BE',
-  },
-  {
-    name: 'Belize',
-    isoAlphaTwoCode: 'BZ',
-  },
-  {
-    name: 'Benin',
-    isoAlphaTwoCode: 'BJ',
-  },
-  {
-    name: 'Bermuda',
-    isoAlphaTwoCode: 'BM',
-  },
-  {
-    name: 'Bhutan',
-    isoAlphaTwoCode: 'BT',
-  },
-  {
-    name: 'Bolivia',
-    isoAlphaTwoCode: 'BO',
-  },
-  {
-    name: 'Bonaire, Sint Eustatius and Saba',
-    isoAlphaTwoCode: 'BQ',
-  },
-  {
-    name: 'Bosnia and Herzegovina',
-    isoAlphaTwoCode: 'BA',
-  },
-  {
-    name: 'Botswana',
-    isoAlphaTwoCode: 'BW',
-  },
-  {
-    name: 'Bouvet Island',
-    isoAlphaTwoCode: 'BV',
-  },
-  {
-    name: 'Brazil',
-    isoAlphaTwoCode: 'BR',
-  },
-  {
-    name: 'British Indian Ocean Territory',
-    isoAlphaTwoCode: 'IO',
-  },
-  {
-    name: 'Brunei',
-    isoAlphaTwoCode: 'BN',
-  },
-  {
-    name: 'Bulgaria',
-    isoAlphaTwoCode: 'BG',
-  },
-  {
-    name: 'Burkina Faso',
-    isoAlphaTwoCode: 'BF',
-  },
-  {
-    name: 'Burundi',
-    isoAlphaTwoCode: 'BI',
-  },
-  {
-    name: 'Cabo Verde',
-    isoAlphaTwoCode: 'CV',
-  },
-  {
-    name: 'Cambodia',
-    isoAlphaTwoCode: 'KH',
-  },
-  {
-    name: 'Cameroon',
-    isoAlphaTwoCode: 'CM',
-  },
-  {
-    name: 'Canada',
-    isoAlphaTwoCode: 'CA',
-  },
-  {
-    name: 'Cayman Islands',
-    isoAlphaTwoCode: 'KY',
-  },
-  {
-    name: 'Central African Republic',
-    isoAlphaTwoCode: 'CF',
-  },
-  {
-    name: 'Chad',
-    isoAlphaTwoCode: 'TD',
-  },
-  {
-    name: 'Chile',
-    isoAlphaTwoCode: 'CL',
-  },
-  {
-    name: 'China',
-    isoAlphaTwoCode: 'CN',
-  },
-  {
-    name: 'Christmas Island',
-    isoAlphaTwoCode: 'CX',
-  },
-  {
-    name: 'Cocos (Keeling) Islands',
-    isoAlphaTwoCode: 'CC',
-  },
-  {
-    name: 'Colombia',
-    isoAlphaTwoCode: 'CO',
-  },
-  {
-    name: 'Comoros',
-    isoAlphaTwoCode: 'KM',
-  },
-  {
-    name: 'Congo',
-    isoAlphaTwoCode: 'CG',
-  },
-  {
-    name: 'Congo (the Democratic Republic of the)',
-    isoAlphaTwoCode: 'CD',
-  },
-  {
-    name: 'Cook Islands',
-    isoAlphaTwoCode: 'CK',
-  },
-  {
-    name: 'Costa Rica',
-    isoAlphaTwoCode: 'CR',
-  },
-  {
-    name: 'Côte dIvoire',
-    isoAlphaTwoCode: 'CI',
-  },
-  {
-    name: 'Croatia',
-    isoAlphaTwoCode: 'HR',
-  },
-  {
-    name: 'Cuba',
-    isoAlphaTwoCode: 'CU',
-  },
-  {
-    name: 'Curaçao',
-    isoAlphaTwoCode: 'CW',
-  },
-  {
-    name: 'Cyprus',
-    isoAlphaTwoCode: 'CY',
-  },
-  {
-    name: 'Czechia',
-    isoAlphaTwoCode: 'CZ',
-  },
-  {
-    name: 'Denmark',
-    isoAlphaTwoCode: 'DK',
-  },
-  {
-    name: 'Djibouti',
-    isoAlphaTwoCode: 'DJ',
-  },
-  {
-    name: 'Dominica',
-    isoAlphaTwoCode: 'DM',
-  },
-  {
-    name: 'Dominican Republic',
-    isoAlphaTwoCode: 'DO',
-  },
-  {
-    name: 'Ecuador',
-    isoAlphaTwoCode: 'EC',
-  },
-  {
-    name: 'Egypt',
-    isoAlphaTwoCode: 'EG',
-  },
-  {
-    name: 'El Salvador',
-    isoAlphaTwoCode: 'SV',
-  },
-  {
-    name: 'Equatorial Guinea',
-    isoAlphaTwoCode: 'GQ',
-  },
-  {
-    name: 'Eritrea',
-    isoAlphaTwoCode: 'ER',
-  },
-  {
-    name: 'Estonia',
-    isoAlphaTwoCode: 'EE',
-  },
-  {
-    name: 'Eswatini',
-    isoAlphaTwoCode: 'SZ',
-  },
-  {
-    name: 'Ethiopia',
-    isoAlphaTwoCode: 'ET',
-  },
-  {
-    name: 'Falkland Islands  [Malvinas]',
-    isoAlphaTwoCode: 'FK',
-  },
-  {
-    name: 'Faroe Islands',
-    isoAlphaTwoCode: 'FO',
-  },
-  {
-    name: 'Fiji',
-    isoAlphaTwoCode: 'FJ',
-  },
-  {
-    name: 'Finland',
-    isoAlphaTwoCode: 'FI',
-  },
-  {
-    name: 'France',
-    isoAlphaTwoCode: 'FR',
-  },
-  {
-    name: 'French Guiana',
-    isoAlphaTwoCode: 'GF',
-  },
-  {
-    name: 'French Polynesia',
-    isoAlphaTwoCode: 'PF',
-  },
-  {
-    name: 'French Southern Territories',
-    isoAlphaTwoCode: 'TF',
-  },
-  {
-    name: 'Gabon',
-    isoAlphaTwoCode: 'GA',
-  },
-  {
-    name: 'Gambia',
-    isoAlphaTwoCode: 'GM',
-  },
-  {
-    name: 'Georgia',
-    isoAlphaTwoCode: 'GE',
-  },
-  {
-    name: 'Germany',
-    isoAlphaTwoCode: 'DE',
-  },
-  {
-    name: 'Ghana',
-    isoAlphaTwoCode: 'GH',
-  },
-  {
-    name: 'Gibraltar',
-    isoAlphaTwoCode: 'GI',
-  },
-  {
-    name: 'Greece',
-    isoAlphaTwoCode: 'GR',
-  },
-  {
-    name: 'Greenland',
-    isoAlphaTwoCode: 'GL',
-  },
-  {
-    name: 'Grenada',
-    isoAlphaTwoCode: 'GD',
-  },
-  {
-    name: 'Guadeloupe',
-    isoAlphaTwoCode: 'GP',
-  },
-  {
-    name: 'Guam',
-    isoAlphaTwoCode: 'GU',
-  },
-  {
-    name: 'Guatemala',
-    isoAlphaTwoCode: 'GT',
-  },
-  {
-    name: 'Guernsey',
-    isoAlphaTwoCode: 'GG',
-  },
-  {
-    name: 'Guinea',
-    isoAlphaTwoCode: 'GN',
-  },
-  {
-    name: 'Guinea-Bissau',
-    isoAlphaTwoCode: 'GW',
-  },
-  {
-    name: 'Guyana',
-    isoAlphaTwoCode: 'GY',
-  },
-  {
-    name: 'Haiti',
-    isoAlphaTwoCode: 'HT',
-  },
-  {
-    name: 'Heard Island and McDonald Islands',
-    isoAlphaTwoCode: 'HM',
-  },
-  {
-    name: 'Holy See',
-    isoAlphaTwoCode: 'VA',
-  },
-  {
-    name: 'Honduras',
-    isoAlphaTwoCode: 'HN',
-  },
-  {
-    name: 'Hong Kong',
-    isoAlphaTwoCode: 'HK',
-  },
-  {
-    name: 'Hungary',
-    isoAlphaTwoCode: 'HU',
-  },
-  {
-    name: 'Iceland',
-    isoAlphaTwoCode: 'IS',
-  },
-  {
-    name: 'India',
-    isoAlphaTwoCode: 'IN',
-  },
-  {
-    name: 'Indonesia',
-    isoAlphaTwoCode: 'ID',
-  },
-  {
-    name: 'Iran',
-    isoAlphaTwoCode: 'IR',
-  },
-  {
-    name: 'Iraq',
-    isoAlphaTwoCode: 'IQ',
-  },
-  {
-    name: 'Ireland',
-    isoAlphaTwoCode: 'IE',
-  },
-  {
-    name: 'Isle of Man',
-    isoAlphaTwoCode: 'IM',
-  },
-  {
-    name: 'Israel',
-    isoAlphaTwoCode: 'IL',
-  },
-  {
-    name: 'Italy',
-    isoAlphaTwoCode: 'IT',
-  },
-  {
-    name: 'Jamaica',
-    isoAlphaTwoCode: 'JM',
-  },
-  {
-    name: 'Japan',
-    isoAlphaTwoCode: 'JP',
-  },
-  {
-    name: 'Jersey',
-    isoAlphaTwoCode: 'JE',
-  },
-  {
-    name: 'Jordan',
-    isoAlphaTwoCode: 'JO',
-  },
-  {
-    name: 'Kazakhstan',
-    isoAlphaTwoCode: 'KZ',
-  },
-  {
-    name: 'Kenya',
-    isoAlphaTwoCode: 'KE',
-  },
-  {
-    name: 'Kiribati',
-    isoAlphaTwoCode: 'KI',
-  },
-  {
-    name: 'Kuwait',
-    isoAlphaTwoCode: 'KW',
-  },
-  {
-    name: 'Kyrgyzstan',
-    isoAlphaTwoCode: 'KG',
-  },
-  {
-    name: 'Laos',
-    isoAlphaTwoCode: 'LA',
-  },
-  {
-    name: 'Latvia',
-    isoAlphaTwoCode: 'LV',
-  },
-  {
-    name: 'Lebanon',
-    isoAlphaTwoCode: 'LB',
-  },
-  {
-    name: 'Lesotho',
-    isoAlphaTwoCode: 'LS',
-  },
-  {
-    name: 'Liberia',
-    isoAlphaTwoCode: 'LR',
-  },
-  {
-    name: 'Libya',
-    isoAlphaTwoCode: 'LY',
-  },
-  {
-    name: 'Liechtenstein',
-    isoAlphaTwoCode: 'LI',
-  },
-  {
-    name: 'Lithuania',
-    isoAlphaTwoCode: 'LT',
-  },
-  {
-    name: 'Luxembourg',
-    isoAlphaTwoCode: 'LU',
-  },
-  {
-    name: 'Macao',
-    isoAlphaTwoCode: 'MO',
-  },
-  {
-    name: 'Madagascar',
-    isoAlphaTwoCode: 'MG',
-  },
-  {
-    name: 'Malawi',
-    isoAlphaTwoCode: 'MW',
-  },
-  {
-    name: 'Malaysia',
-    isoAlphaTwoCode: 'MY',
-  },
-  {
-    name: 'Maldives',
-    isoAlphaTwoCode: 'MV',
-  },
-  {
-    name: 'Mali',
-    isoAlphaTwoCode: 'ML',
-  },
-  {
-    name: 'Malta',
-    isoAlphaTwoCode: 'MT',
-  },
-  {
-    name: 'Marshall Islands',
-    isoAlphaTwoCode: 'MH',
-  },
-  {
-    name: 'Martinique',
-    isoAlphaTwoCode: 'MQ',
-  },
-  {
-    name: 'Mauritania',
-    isoAlphaTwoCode: 'MR',
-  },
-  {
-    name: 'Mauritius',
-    isoAlphaTwoCode: 'MU',
-  },
-  {
-    name: 'Mayotte',
-    isoAlphaTwoCode: 'YT',
-  },
-  {
-    name: 'Mexico',
-    isoAlphaTwoCode: 'MX',
-  },
-  {
-    name: 'Micronesia (Federated States of)',
-    isoAlphaTwoCode: 'FM',
-  },
-  {
-    name: 'Moldova',
-    isoAlphaTwoCode: 'MD',
-  },
-  {
-    name: 'Monaco',
-    isoAlphaTwoCode: 'MC',
-  },
-  {
-    name: 'Mongolia',
-    isoAlphaTwoCode: 'MN',
-  },
-  {
-    name: 'Montenegro',
-    isoAlphaTwoCode: 'ME',
-  },
-  {
-    name: 'Montserrat',
-    isoAlphaTwoCode: 'MS',
-  },
-  {
-    name: 'Morocco',
-    isoAlphaTwoCode: 'MA',
-  },
-  {
-    name: 'Mozambique',
-    isoAlphaTwoCode: 'MZ',
-  },
-  {
-    name: 'Myanmar',
-    isoAlphaTwoCode: 'MM',
-  },
-  {
-    name: 'Namibia',
-    isoAlphaTwoCode: 'NA',
-  },
-  {
-    name: 'Nauru',
-    isoAlphaTwoCode: 'NR',
-  },
-  {
-    name: 'Nepal',
-    isoAlphaTwoCode: 'NP',
-  },
-  {
-    name: 'Netherlands',
-    isoAlphaTwoCode: 'NL',
-  },
-  {
-    name: 'New Caledonia',
-    isoAlphaTwoCode: 'NC',
-  },
-  {
-    name: 'New Zealand',
-    isoAlphaTwoCode: 'NZ',
-  },
-  {
-    name: 'Nicaragua',
-    isoAlphaTwoCode: 'NI',
-  },
-  {
-    name: 'Niger',
-    isoAlphaTwoCode: 'NE',
-  },
-  {
-    name: 'Nigeria',
-    isoAlphaTwoCode: 'NG',
-  },
-  {
-    name: 'Niue',
-    isoAlphaTwoCode: 'NU',
-  },
-  {
-    name: 'Norfolk Island',
-    isoAlphaTwoCode: 'NF',
-  },
-  {
-    name: 'North Korea',
-    isoAlphaTwoCode: 'KP',
-  },
-  {
-    name: 'North Macedonia',
-    isoAlphaTwoCode: 'MK',
-  },
-  {
-    name: 'Northern Mariana Islands',
-    isoAlphaTwoCode: 'MP',
-  },
-  {
-    name: 'Norway',
-    isoAlphaTwoCode: 'NO',
-  },
-  {
-    name: 'Oman',
-    isoAlphaTwoCode: 'OM',
-  },
-  {
-    name: 'Pakistan',
-    isoAlphaTwoCode: 'PK',
-  },
-  {
-    name: 'Palau',
-    isoAlphaTwoCode: 'PW',
-  },
-  {
-    name: 'Palestine, State of',
-    isoAlphaTwoCode: 'PS',
-  },
-  {
-    name: 'Panama',
-    isoAlphaTwoCode: 'PA',
-  },
-  {
-    name: 'Papua New Guinea',
-    isoAlphaTwoCode: 'PG',
-  },
-  {
-    name: 'Paraguay',
-    isoAlphaTwoCode: 'PY',
-  },
-  {
-    name: 'Peru',
-    isoAlphaTwoCode: 'PE',
-  },
-  {
-    name: 'Philippines',
-    isoAlphaTwoCode: 'PH',
-  },
-  {
-    name: 'Pitcairn',
-    isoAlphaTwoCode: 'PN',
-  },
-  {
-    name: 'Poland',
-    isoAlphaTwoCode: 'PL',
-  },
-  {
-    name: 'Portugal',
-    isoAlphaTwoCode: 'PT',
-  },
-  {
-    name: 'Puerto Rico',
-    isoAlphaTwoCode: 'PR',
-  },
-  {
-    name: 'Qatar',
-    isoAlphaTwoCode: 'QA',
-  },
-  {
-    name: 'Réunion',
-    isoAlphaTwoCode: 'RE',
-  },
-  {
-    name: 'Romania',
-    isoAlphaTwoCode: 'RO',
-  },
-  {
-    name: 'Russia',
-    isoAlphaTwoCode: 'RU',
-  },
-  {
-    name: 'Rwanda',
-    isoAlphaTwoCode: 'RW',
-  },
-  {
-    name: 'Saint Barthélemy',
-    isoAlphaTwoCode: 'BL',
-  },
-  {
-    name: 'Saint Helena, Ascension and Tristan da Cunha',
-    isoAlphaTwoCode: 'SH',
-  },
-  {
-    name: 'Saint Kitts and Nevis',
-    isoAlphaTwoCode: 'KN',
-  },
-  {
-    name: 'Saint Lucia',
-    isoAlphaTwoCode: 'LC',
-  },
-  {
-    name: 'Saint Martin (French part)',
-    isoAlphaTwoCode: 'MF',
-  },
-  {
-    name: 'Saint Pierre and Miquelon',
-    isoAlphaTwoCode: 'PM',
-  },
-  {
-    name: 'Saint Vincent and the Grenadines',
-    isoAlphaTwoCode: 'VC',
-  },
-  {
-    name: 'Samoa',
-    isoAlphaTwoCode: 'WS',
-  },
-  {
-    name: 'San Marino',
-    isoAlphaTwoCode: 'SM',
-  },
-  {
-    name: 'Sao Tome and Principe',
-    isoAlphaTwoCode: 'ST',
-  },
-  {
-    name: 'Saudi Arabia',
-    isoAlphaTwoCode: 'SA',
-  },
-  {
-    name: 'Senegal',
-    isoAlphaTwoCode: 'SN',
-  },
-  {
-    name: 'Serbia',
-    isoAlphaTwoCode: 'RS',
-  },
-  {
-    name: 'Seychelles',
-    isoAlphaTwoCode: 'SC',
-  },
-  {
-    name: 'Sierra Leone',
-    isoAlphaTwoCode: 'SL',
-  },
-  {
-    name: 'Singapore',
-    isoAlphaTwoCode: 'SG',
-  },
-  {
-    name: 'Sint Maarten (Dutch part)',
-    isoAlphaTwoCode: 'SX',
-  },
-  {
-    name: 'Slovakia',
-    isoAlphaTwoCode: 'SK',
-  },
-  {
-    name: 'Slovenia',
-    isoAlphaTwoCode: 'SI',
-  },
-  {
-    name: 'Solomon Islands',
-    isoAlphaTwoCode: 'SB',
-  },
-  {
-    name: 'Somalia',
-    isoAlphaTwoCode: 'SO',
-  },
-  {
-    name: 'South Africa',
-    isoAlphaTwoCode: 'ZA',
-  },
-  {
-    name: 'South Georgia and the South Sandwich Islands',
-    isoAlphaTwoCode: 'GS',
-  },
-  {
-    name: 'South Korea',
-    isoAlphaTwoCode: 'KR',
-  },
-  {
-    name: 'South Sudan',
-    isoAlphaTwoCode: 'SS',
-  },
-  {
-    name: 'Spain',
-    isoAlphaTwoCode: 'ES',
-  },
-  {
-    name: 'Sri Lanka',
-    isoAlphaTwoCode: 'LK',
-  },
-  {
-    name: 'Sudan',
-    isoAlphaTwoCode: 'SD',
-  },
-  {
-    name: 'Suriname',
-    isoAlphaTwoCode: 'SR',
-  },
-  {
-    name: 'Svalbard and Jan Mayen',
-    isoAlphaTwoCode: 'SJ',
-  },
-  {
-    name: 'Sweden',
-    isoAlphaTwoCode: 'SE',
-  },
-  {
-    name: 'Switzerland',
-    isoAlphaTwoCode: 'CH',
-  },
-  {
-    name: 'Syria',
-    isoAlphaTwoCode: 'SY',
-  },
-  {
-    name: 'Taiwan',
-    isoAlphaTwoCode: 'TW',
-  },
-  {
-    name: 'Tajikistan',
-    isoAlphaTwoCode: 'TJ',
-  },
-  {
-    name: 'Tanzania',
-    isoAlphaTwoCode: 'TZ',
-  },
-  {
-    name: 'Thailand',
-    isoAlphaTwoCode: 'TH',
-  },
-  {
-    name: 'Timor-Leste',
-    isoAlphaTwoCode: 'TL',
-  },
-  {
-    name: 'Togo',
-    isoAlphaTwoCode: 'TG',
-  },
-  {
-    name: 'Tokelau',
-    isoAlphaTwoCode: 'TK',
-  },
-  {
-    name: 'Tonga',
-    isoAlphaTwoCode: 'TO',
-  },
-  {
-    name: 'Trinidad and Tobago',
-    isoAlphaTwoCode: 'TT',
-  },
-  {
-    name: 'Tunisia',
-    isoAlphaTwoCode: 'TN',
-  },
-  {
-    name: 'Turkey',
-    isoAlphaTwoCode: 'TR',
-  },
-  {
-    name: 'Turkmenistan',
-    isoAlphaTwoCode: 'TM',
-  },
-  {
-    name: 'Turks and Caicos Islands',
-    isoAlphaTwoCode: 'TC',
-  },
-  {
-    name: 'Tuvalu',
-    isoAlphaTwoCode: 'TV',
-  },
-  {
-    name: 'Uganda',
-    isoAlphaTwoCode: 'UG',
-  },
-  {
-    name: 'Ukraine',
-    isoAlphaTwoCode: 'UA',
-  },
-  {
-    name: 'United Arab Emirates',
-    isoAlphaTwoCode: 'AE',
-  },
-  {
-    name: 'United Kingdom',
-    isoAlphaTwoCode: 'GB',
-  },
-  {
-    name: 'United States Minor Outlying Islands',
-    isoAlphaTwoCode: 'UM',
-  },
-  {
-    name: 'United States of America',
-    isoAlphaTwoCode: 'US',
-  },
-  {
-    name: 'Uruguay',
-    isoAlphaTwoCode: 'UY',
-  },
-  {
-    name: 'Uzbekistan',
-    isoAlphaTwoCode: 'UZ',
-  },
-  {
-    name: 'Vanuatu',
-    isoAlphaTwoCode: 'VU',
-  },
-  {
-    name: 'Venezuela',
-    isoAlphaTwoCode: 'VE',
-  },
-  {
-    name: 'Vietnam',
-    isoAlphaTwoCode: 'VN',
-  },
-  {
-    name: 'Virgin Islands (British)',
-    isoAlphaTwoCode: 'VG',
-  },
-  {
-    name: 'Virgin Islands (U.S.)',
-    isoAlphaTwoCode: 'VI',
-  },
-  {
-    name: 'Wallis and Futuna',
-    isoAlphaTwoCode: 'WF',
-  },
-  {
-    name: 'Western Sahara',
-    isoAlphaTwoCode: 'EH',
-  },
-  {
-    name: 'Yemen',
-    isoAlphaTwoCode: 'YE',
-  },
-  {
-    name: 'Zambia',
-    isoAlphaTwoCode: 'ZM',
-  },
-  {
-    name: 'Zimbabwe',
-    isoAlphaTwoCode: 'ZW',
-  },
-];
